# tasks/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  dto.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  dto.TaskRequest:
    properties:
      description:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  dto.TaskResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.UpdateTaskRequest:
    properties:
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: fvi-it@mail.ru
    name: Идель Фазлетдинов
  description: Сервис для создания Tasks.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Tasks Service
  version: "1.0"
paths:
  /task:
    post:
      consumes:
      - application/json
      parameters:
      - description: Создание Task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Создание Task
      tags:
      - Tasks
  /task/{task_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Удаление Task
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получение Task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      - description: Для обновления Task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Обновление Task
      tags:
      - Tasks
  /tasks:
    get:
      consumes:
      - application/json
      parameters:
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: offset
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TaskResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получение списока Task
      tags:
      - Tasks
swagger: "2.0"